requirement_modeling_project/requirements.txt
requests==2.32.4
openai==1.88.0
python-dotenv==1.1.0
pydantic==2.11.7
fastapi==0.115.6
uvicorn==0.34.3 